Fix for updating dirty state of render boxes when re-attached
